Move test code outside of src
