+++ test/runTest.ts
@@ -20,7 +20,7 @@ async function main() {
     try {
         // The folder containing the Extension Manifest package.json
         // Passed to `--extensionDevelopmentPath`
-        const extensionDevelopmentPath = path.resolve(__dirname, "../../");
+        const extensionDevelopmentPath = path.resolve(__dirname, "../");
 
         // The path to test runner
         // Passed to --extensionTestsPath
