+++ tsconfig.json
@@ -7,7 +7,6 @@
 			"ES2020"
 		],
 		"sourceMap": true,
-		"rootDir": "src",
 		"strict": true   /* enable all strict type-checking options */
 		/* Additional Checks */
 		// "noImplicitReturns": true, /* Report error when not all code paths in function return a value. */
